Fixed the quoted NULL value problem.  The code wasn't checking for null values explicitly when building query strings, it was only checking the attributes of the data type.  That is, it might ask something like, "Should sql_text_null be quoted", get "yes", and quote the value when building SQL query strings even when the value is a SQL NULL, and thus needs to go literally into the generated SQL.
MySQL++ must have gotten away without such code by playing some kind of
game with the field type.  This is a better way to do it.

Incidentally, this marks the first version of this branch that runs dtest
to completion without diffs!  A bit of cleanup, and we can merge it back
into the trunk!

FossilOrigin-Name: e32649510e044da99bab216c0448f8dbcae76fc3